requirements.txt
aiohttp==3.6.2
alembic==1.4.2
amqp==2.6.1
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
blis==0.4.1
bs4==0.0.1
catalogue==1.0.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
cymem==2.0.3
decorator==4.4.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
kombu==4.6.11
Mako==1.1.3
MarkupSafe==1.1.1
multidict==4.7.6
murmurhash==1.0.2
nltk==3.5
numpy==1.19.0
pandas==1.0.5
plac==1.1.3
preshed==3.0.2
psycopg2-binary==2.8.5
pycparser==2.20
python-dateutil==2.8.1
python-editor==1.0.4
python-telegram-bot==12.8
pytz==2020.1
redis==3.5.3
regex==2020.6.8
requests==2.24.0
scikit-learn==0.23.1
scipy==1.5.0
six==1.15.0
soupsieve==2.0.1
spacy==2.3.0
SQLAlchemy==1.3.18
srsly==1.0.2
thinc==7.4.1
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.47.0
urllib3==1.25.9
vine==1.3.0
wasabi==0.7.0
Werkzeug==1.0.1
yarl==1.4.2
zipp==3.1.0
